pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "leticia", email = "leticiaborgess.oficial@gmail.com"},
]
dependencies = [
    "django>=4.2.5",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=4.2.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

